7th March

+++ src/app/admin/users/users.page.ts
@@ -1,6 +1,7 @@
 import { ResourceLoader } from '@angular/compiler';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { AlertController } from '@ionic/angular';
 import { UserService } from 'src/app/user.service';
 
 @Component({
@@ -10,19 +11,45 @@ import { UserService } from 'src/app/user.service';
 })
 export class UsersPage implements OnInit {
   users:any
-  constructor(private userService:UserService, private route:ActivatedRoute) { }
+  constructor(private userService:UserService, private route:ActivatedRoute, public alertCtrl:AlertController) { }
 
   ngOnInit() {
-      this.userService.getUsers().subscribe((res)=>{
-        console.log(res)
-        this.users = res
-      });
+      // this.userService.getUsers().subscribe((res)=>{
+      //   console.log(res)
+      //   this.users = res
+      // });
+  }
+  ionViewWillEnter(){
+    this.userService.getUsers().subscribe((res)=>{
+      console.log(res)
+      this.users = res
+    })
   }
 
-  deleteUser(id){
-    this.userService.deleteUser(id).subscribe((res)=>{
-      console.log(res)
-    })
+  async deleteUser(id){
+    const confirm = await this.alertCtrl.create({
+      header: 'Are you sure?',
+      message: 'Do you want to delete this user?',
+      buttons: [
+        {
+          text: 'Yes',
+          handler: () => {
+            this.userService.deleteUser(id).subscribe((res) => {
+              console.log(res);
+            });
+            this.ionViewWillEnter();
+          }
+        },
+        {
+          text: 'No',
+          handler: () => {
+            console.log("Close");
+          }
+        }
+      ]
+    });
+    await confirm.present();
+
   }
 
 }
